api: Create separate session table
