+++ packages/api/src/webhooks/cannon.ts
@@ -4,7 +4,8 @@ import isLocalIP from "is-local-ip";
 import { Response } from "node-fetch";
 import { v4 as uuid } from "uuid";
 import { parse as parseUrl } from "url";
-import { DB, DBSession } from "../store/db";
+import { DB } from "../store/db";
+import { DBSession } from "../store/session-table";
 import messages from "../store/messages";
 import Queue from "../store/queue";
 import { DBWebhook } from "../store/webhook-table";
